requirements.txt
colorama
scapy
requests
pycryptodome 
